牌桌筹码数据类型修改

+++ src/main/java/com/dx/facade/venue/dto/DxTableInfoDto.java
@@ -64,10 +64,10 @@ public class DxTableInfoDto implements Serializable {
     private BigDecimal bringOutChipScore;
 
     @ApiModelProperty("小盲注")
-    private Long sbBlindScore;
+    private BigDecimal sbBlindScore;
 
     @ApiModelProperty("大盲注")
-    private Long bbBlindScore;
+    private BigDecimal bbBlindScore;
 
     @ApiModelProperty("抽水比例")
     private BigDecimal rake;
