+++ src/main/java/com/dx/facade/venue/req/DxTableInfoAddReq.java
@@ -71,10 +71,10 @@ public class DxTableInfoAddReq implements Serializable {
     private BigDecimal bringOutChipScore;
 
     @ApiModelProperty("小盲注")
-    private Long sbBlindScore;
+    private BigDecimal sbBlindScore;
 
     @ApiModelProperty("大盲注")
-    private Long bbBlindScore;
+    private BigDecimal bbBlindScore;
 
     @ApiModelProperty("抽水比例")
     private BigDecimal rake;
