+++ src/main/java/com/dx/facade/venue/resp/TableConfigRuleResp.java
@@ -45,17 +45,17 @@ public class TableConfigRuleResp {
     /**
      * 默认带入的筹码
      */
-    private Long bringChipScore;
+    private BigDecimal bringChipScore;
 
     /***
      * 小盲注
      */
-    private Long sbBlindScore;
+    private BigDecimal sbBlindScore;
 
     /***
      * 大盲注
      */
-    private Long bbBlindScore;
+    private BigDecimal bbBlindScore;
 
     /**
      * 时长 单位 分钟
@@ -65,12 +65,12 @@ public class TableConfigRuleResp {
     /**
      * 最小带入的筹码
      */
-    private Long minBringChipScore;
+    private BigDecimal minBringChipScore;
 
     /**
      * 最大带入的筹码
      */
-    private Long maxBringChipScore;
+    private BigDecimal maxBringChipScore;
 
     /**
      * 抽水比例配置
